refactor: extract manifest-harness codegen from cli.rs into its own module
Pure move: HarnessOptions/HarnessMode, run_manifest_harness, the inline
harness Cargo.toml/main.rs templates, cargo-metadata package resolution,
and entrypoint validation now live in src/manifest_harness.rs beside
generate/, with their unit tests. cli.rs keeps the clap surface and the
scaffold side effects; the shared path helpers it owns are pub(crate) for
the harness. No behavior change.
